% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\name{bg_data}
\alias{bg_data}
\title{bg_data}
\usage{
bg_data(geoid = NULL)
}
\arguments{
\item{geoid}{c(x,y,z) a vector of block group IDs to return}
}
\description{
Load Census block group data for Minnesota.
}
\examples{
# All block groups
bgs <- bg_data()

# A single block group
bgs <- bg_data(geoid = c(271090017024))
}
\keyword{block}
\keyword{blockgroups}
\keyword{census}
\keyword{group}
