% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_shape.R
\name{bg_shape}
\alias{bg_shape}
\title{bg_shape}
\usage{
bg_shape(geoid = NULL)
}
\arguments{
\item{geoid}{c(x,y,z) a vector of block group IDs to return}
}
\description{
Load Census block group shapefile for Minnesota.
}
\examples{
# All block groups
bgs <- bg_shape()

# A single block group
bgs <- bg_shape(geoid = c(271090017024))
}
\keyword{block}
\keyword{blockgroups}
\keyword{census}
\keyword{group}
