% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_data.R
\name{get_blockgroups}
\alias{get_blockgroups}
\title{get_blockgroups}
\usage{
get_blockgroups(geoid = NULL)
}
\arguments{
\item{geoid}{c(x,y,z) a vector of block group IDs to return}
}
\description{
Load Census block group data for Minnesota.
}
\examples{
# All block groups
bgs <- get_blockgroups()

# A single block group
bgs <- get_blockgroups(geoid = c(271090017024))
}
\keyword{block}
\keyword{blockgroups}
\keyword{census}
\keyword{group}
