% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_receptors.R, R/receptors.R
\name{get_receptors}
\alias{get_receptors}
\alias{get_receptors}
\title{get_receptors}
\usage{
get_receptors(bg_id = NULL)

get_receptors(bg_id = NULL)
}
\arguments{
\item{bg_id}{Block group IDs to include. Default includes all.}

\item{bg_id}{Block group IDs to include. Default includes all.}
}
\description{
MNRISKS 2011 modeled receptors

MNRISKS 2011 modeled receptors
}
\examples{
# All receptors
points <- get_receptors()

#For selected block groups
points <- get_receptors(bg_id = c(271090017012))
# All receptors
points <- get_receptors()

#For selected block groups
points <- get_receptors(bg_id = c(271090017012))
}
\keyword{mnrisks}
\keyword{receptors}
