% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_bg_avg.R
\name{idw_bg_avg}
\alias{idw_bg_avg}
\title{idw_bg_avg}
\usage{
idw_bg_avg(data = df, value = "cancer_risk", lat = "Lat", long = "Long",
  bg_id = NULL, metro = FALSE)
}
\arguments{
\item{data}{Data frame with lat / long coordinates and values for averaging.}
}
\description{
Find the IDW spatial block group average for the given receptors using inverse distance weighting
}
\examples{
# For all block groups
bg_avg <- idw_bg_avg(data   = df, 
                     value  = cancer_risk,
                     lat    = lat, 
                     long   = long,
                     bg_id  = NULL)
}
\keyword{average}
\keyword{block}
\keyword{group}
\keyword{mnrisks}
\keyword{receptors}
\keyword{spatial}
