% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw_receptors.R
\name{idw_receptors}
\alias{idw_receptors}
\title{idw_receptors}
\usage{
idw_receptors(column = "cancer_risk", bg_id = NULL)
}
\arguments{
\item{column}{Column name containing the point values to interpolate.}

\item{bg_id}{Block group IDs to include. Default includes all.}
}
\description{
Inverse distance weighting for MNRISKS receptors
}
\examples{
# For all block groups
idw_receptors()

For selected block groups
idw_receptors(column = "cancer_risk", 
              bg_id = c(271090017012))

}
\keyword{grid}
\keyword{polygons}
\keyword{spatial}
\keyword{voronoi}
