% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_interp_grids.R
\name{make_interp_grids}
\alias{make_interp_grids}
\title{make_interp_grids}
\usage{
make_interp_grids(grid_space = 0.00135)
}
\arguments{
\item{grid_space}{Spacing between grid points.}
}
\description{
Create Minnesota interpolation grids
}
\examples{
# For all counties
make_interp_grids()

}
\keyword{grid}
\keyword{receptors}
\keyword{spatial}
