% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_voronoi_polys.R
\name{make_voronoi_polys}
\alias{make_voronoi_polys}
\title{make_voronoi_polys}
\usage{
make_voronoi_polys()
}
\description{
Create Minnesota voronoi grid from receptors
}
\examples{
# For all counties
make_voronoi_polys()

}
\keyword{grid}
\keyword{polygons}
\keyword{spatial}
\keyword{voronoi}
