% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rsk.R
\name{read_rsk}
\alias{read_rsk}
\title{read_rsk}
\usage{
read_rsk(file, scenarios = NULL)
}
\arguments{
\item{file}{Path to .rsk file}
}
\description{
MNRISKS 2011 modeled receptors
}
\examples{
\dontrun {
# Read in a single .rsk file
acrolein <- read_rsk("acrolein_resuls.rsk")
}
}
\keyword{mnrisks}
\keyword{results}
\keyword{rsk}
