% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptor_bg_areas.R
\name{receptor_bg_areas}
\alias{receptor_bg_areas}
\title{receptor_bg_areas}
\usage{
receptor_bg_areas(bg_id = NULL)
}
\arguments{
\item{bg_id}{Block group IDs to include. Default includes all.}
}
\description{
MNRISKS 2011 receptor area coverage for each block gorup
}
\examples{
# All receptors
rec_fractions <- receptor_bg_areas()

#For selected block groups
rec_fractions <- receptor_bg_areas(bg_id = c(271090017012))
}
\keyword{mnrisks}
\keyword{receptors}
