% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/receptors.R
\name{receptors}
\alias{receptors}
\title{receptors}
\usage{
receptors(bg_id = NULL)
}
\arguments{
\item{bg_id}{Block group IDs to include. Default includes all.}
}
\description{
MNRISKS 2011 modeled receptors
}
\examples{
# All receptors
points <- receptors()

#For selected block groups
points <- receptors(bg_id = c(271090017012))
}
\keyword{mnrisks}
\keyword{receptors}
