% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_bg_avg.R
\name{spatial_bg_avg}
\alias{spatial_bg_avg}
\title{spatial_bg_avg}
\usage{
spatial_bg_avg(data = df, values = NULL, receptor = "receptor")
}
\arguments{
\item{data}{Data frame with lat / long coordinates and values for averaging.}
}
\description{
Find the spatial block group average, weighting each receptor by its spatial coverage of the block group
}
\examples{
# For all block groups
bg_avg <- spatial_bg_avg(data     = df, 
                         values   = "cancer_risk",
                         receptor = "receptor"
                         )
}
\keyword{average}
\keyword{block}
\keyword{group}
\keyword{mnrisks}
\keyword{receptors}
\keyword{spatial}
